requirements.txt
huey==2.4.2
PyYAML==6.0
tinydb==4.7.0
typing_extensions==4.1.1
